requirements.txt
# Langchain core
langchain
langchain-core

# openai integration 
langchain-openai
openai

# Anthropic integration
langchain-anthropic
anthropic

# gemini integration
langchain-google-genai
google-generativeai

# hugging face
langchain-huggingface
transformers
huggingface-hub

# RAG and Vector Database
chromadb
langchain-chroma
sentence-transformers

# PDF Processing - FIXED VERSIONS
pypdf
pymupdf
langchain-community

# Web Framework
streamlit
gradio

# envioronmanet variable management
python-dotenv

# machine learning utilities
numpy
scikit-learn
